Add `ComplexPathsGM` test for advanced path rendering with gradients and transformations, corresponding unit test, and register it in `AllTests.kt`. Update Level 2 test progress in `SKIA_DM_TESTS_TO_IMPLEMENT.md`.

+++ kanvas-kotlin/src/test/kotlin/core/TextToGlyphsTest.kt
@@ -158,8 +158,8 @@ class TextToGlyphsTest {
         val largeGlyphRun = largeFont.createGlyphRun(text)
         
         // Larger font should produce larger glyph advances
-        val smallTotalWidth = smallGlyphRun.glyphs.sumOf { it.advanceX }
-        val largeTotalWidth = largeGlyphRun.glyphs.sumOf { it.advanceX }
+        val smallTotalWidth = smallGlyphRun.glyphs.sumOf { it.advanceX.toDouble() }
+        val largeTotalWidth = largeGlyphRun.glyphs.sumOf { it.advanceX.toDouble() }
         
         assertTrue(largeTotalWidth > smallTotalWidth)
     }
